# docs/dist/api.swagger.json
{
  "swagger": "2.0",
  "info": {
    "title": "Railway desktop app - Backend RESTful  API Docs ",
    "version": "1.0",
    "contact": {
      "name": "Railway desktop support team",
      "url": "https://desktop-support.railway-desktop.io",
      "email": "support@railway-desktop.io"
    },
    "license": {
      "name": "Terms and conditions",
      "url": "https://railway-desktop.io/terms"
    }
  },
  "tags": [
    {
      "name": "MybackendGrpcSvc"
    }
  ],
  "schemes": [
    "http",
    "https"
  ],
  "consumes": [
    "application/json"
  ],
  "produces": [
    "application/json"
  ],
  "paths": {
    "/v1/service": {
      "delete": {
        "operationId": "MybackendGrpcSvc_DeleteService",
        "responses": {
          "200": {
            "description": "A successful response.",
            "schema": {
              "$ref": "#/definitions/v1DeleteServiceResponse"
            }
          },
          "default": {
            "description": "An unexpected error response.",
            "schema": {
              "$ref": "#/definitions/rpcStatus"
            }
          }
        },
        "parameters": [
          {
            "name": "serviceId",
            "in": "query",
            "required": false,
            "type": "string"
          }
        ],
        "tags": [
          "MybackendGrpcSvc"
        ]
      },
      "post": {
        "operationId": "MybackendGrpcSvc_ServiceCreateFromImage",
        "responses": {
          "200": {
            "description": "A successful response.",
            "schema": {
              "$ref": "#/definitions/v1ServiceCreateFromImageResponse"
            }
          },
          "default": {
            "description": "An unexpected error response.",
            "schema": {
              "$ref": "#/definitions/rpcStatus"
            }
          }
        },
        "parameters": [
          {
            "name": "body",
            "in": "body",
            "required": true,
            "schema": {
              "$ref": "#/definitions/v1ServiceCreateFromImageRequest"
            }
          }
        ],
        "tags": [
          "MybackendGrpcSvc"
        ]
      }
    },
    "/v1/user/byid": {
      "get": {
        "operationId": "MybackendGrpcSvc_GetUserInfoById",
        "responses": {
          "200": {
            "description": "A successful response.",
            "schema": {
              "$ref": "#/definitions/v1GetUserInfoByIdResponse"
            }
          },
          "default": {
            "description": "An unexpected error response.",
            "schema": {
              "$ref": "#/definitions/rpcStatus"
            }
          }
        },
        "parameters": [
          {
            "name": "userId",
            "in": "query",
            "required": false,
            "type": "string"
          },
          {
            "name": "Authorization",
            "description": "JWT authentication token issued during authentication i.e Bearer \u003c\u003cJWT_TOKEN\u003e\u003e",
            "in": "header",
            "required": true,
            "type": "string"
          }
        ],
        "tags": [
          "MybackendGrpcSvc"
        ]
      }
    },
    "/v1/user/byusername": {
      "get": {
        "operationId": "MybackendGrpcSvc_GetUserInfoByUsername",
        "responses": {
          "200": {
            "description": "A successful response.",
            "schema": {
              "$ref": "#/definitions/v1GetUserInfoByUsernameResponse"
            }
          },
          "default": {
            "description": "An unexpected error response.",
            "schema": {
              "$ref": "#/definitions/rpcStatus"
            }
          }
        },
        "parameters": [
          {
            "name": "userName",
            "in": "query",
            "required": false,
            "type": "string"
          },
          {
            "name": "Authorization",
            "description": "JWT authentication token issued during authentication i.e Bearer \u003c\u003cJWT_TOKEN\u003e\u003e",
            "in": "header",
            "required": true,
            "type": "string"
          }
        ],
        "tags": [
          "MybackendGrpcSvc"
        ]
      }
    }
  },
  "definitions": {
    "protobufAny": {
      "type": "object",
      "properties": {
        "@type": {
          "type": "string",
          "description": "A URL/resource name that uniquely identifies the type of the serialized\nprotocol buffer message. This string must contain at least\none \"/\" character. The last segment of the URL's path must represent\nthe fully qualified name of the type (as in\n`path/google.protobuf.Duration`). The name should be in a canonical form\n(e.g., leading \".\" is not accepted).\n\nIn practice, teams usually precompile into the binary all types that they\nexpect it to use in the context of Any. However, for URLs which use the\nscheme `http`, `https`, or no scheme, one can optionally set up a type\nserver that maps type URLs to message definitions as follows:\n\n* If no scheme is provided, `https` is assumed.\n* An HTTP GET on the URL must yield a [google.protobuf.Type][]\n  value in binary format, or produce an error.\n* Applications are allowed to cache lookup results based on the\n  URL, or have them precompiled into a binary to avoid any\n  lookup. Therefore, binary compatibility needs to be preserved\n  on changes to types. (Use versioned type names to manage\n  breaking changes.)\n\nNote: this functionality is not currently available in the official\nprotobuf release, and it is not used for type URLs beginning with\ntype.googleapis.com. As of May 2023, there are no widely used type server\nimplementations and no plans to implement one.\n\nSchemes other than `http`, `https` (or the empty scheme) might be\nused with implementation specific semantics."
        }
      },
      "additionalProperties": {},
      "description": "`Any` contains an arbitrary serialized protocol buffer message along with a\nURL that describes the type of the serialized message.\n\nProtobuf library provides support to pack/unpack Any values in the form\nof utility functions or additional generated methods of the Any type.\n\nExample 1: Pack and unpack a message in C++.\n\n    Foo foo = ...;\n    Any any;\n    any.PackFrom(foo);\n    ...\n    if (any.UnpackTo(\u0026foo)) {\n      ...\n    }\n\nExample 2: Pack and unpack a message in Java.\n\n    Foo foo = ...;\n    Any any = Any.pack(foo);\n    ...\n    if (any.is(Foo.class)) {\n      foo = any.unpack(Foo.class);\n    }\n    // or ...\n    if (any.isSameTypeAs(Foo.getDefaultInstance())) {\n      foo = any.unpack(Foo.getDefaultInstance());\n    }\n\n Example 3: Pack and unpack a message in Python.\n\n    foo = Foo(...)\n    any = Any()\n    any.Pack(foo)\n    ...\n    if any.Is(Foo.DESCRIPTOR):\n      any.Unpack(foo)\n      ...\n\n Example 4: Pack and unpack a message in Go\n\n     foo := \u0026pb.Foo{...}\n     any, err := anypb.New(foo)\n     if err != nil {\n       ...\n     }\n     ...\n     foo := \u0026pb.Foo{}\n     if err := any.UnmarshalTo(foo); err != nil {\n       ...\n     }\n\nThe pack methods provided by protobuf library will by default use\n'type.googleapis.com/full.type.name' as the type URL and the unpack\nmethods only use the fully qualified type name after the last '/'\nin the type URL, for example \"foo.bar.com/x/y.z\" will yield type\nname \"y.z\".\n\nJSON\n====\nThe JSON representation of an `Any` value uses the regular\nrepresentation of the deserialized, embedded message, with an\nadditional field `@type` which contains the type URL. Example:\n\n    package google.profile;\n    message Person {\n      string first_name = 1;\n      string last_name = 2;\n    }\n\n    {\n      \"@type\": \"type.googleapis.com/google.profile.Person\",\n      \"firstName\": \u003cstring\u003e,\n      \"lastName\": \u003cstring\u003e\n    }\n\nIf the embedded message type is well-known and has a custom JSON\nrepresentation, that representation will be embedded adding a field\n`value` which holds the custom JSON in addition to the `@type`\nfield. Example (for message [google.protobuf.Duration][]):\n\n    {\n      \"@type\": \"type.googleapis.com/google.protobuf.Duration\",\n      \"value\": \"1.212s\"\n    }"
    },
    "rpcStatus": {
      "type": "object",
      "properties": {
        "code": {
          "type": "integer",
          "format": "int32"
        },
        "message": {
          "type": "string"
        },
        "details": {
          "type": "array",
          "items": {
            "type": "object",
            "$ref": "#/definitions/protobufAny"
          }
        }
      }
    },
    "v1ServiceCreateFromImageRequest": {
      "type": "object",
      "properties": {
        "imageUrl": {
          "type": "string"
        },
        "projectId": {
          "type": "string"
        },
        "environmentId": {
          "type": "string"
        }
      }
    },
    "v1ServiceCreateFromImageResponse": {
      "type": "object",
      "properties": {
        "serviceId": {
          "type": "string"
        },
        "error": {
          "$ref": "#/definitions/v1Error"
        }
      }
    },
    "v1DeleteServiceResponse": {
      "type": "object",
      "properties": {
        "serviceId": {
          "type": "string"
        },
        "environmentId": {
          "type": "string"
        },
        "error": {
          "$ref": "#/definitions/v1Error"
        }
      }
    },
    "v1Error": {
      "type": "object",
      "properties": {
        "code": {
          "type": "integer",
          "format": "int64"
        },
        "message": {
          "type": "string"
        }
      },
      "title": "#######"
    },
    "v1GetUserInfoByIdResponse": {
      "type": "object",
      "properties": {
        "error": {
          "$ref": "#/definitions/v1Error"
        },
        "result": {
          "$ref": "#/definitions/v1GetUserInfoByIdResult"
        }
      }
    },
    "v1GetUserInfoByIdResult": {
      "type": "object",
      "properties": {
        "userId": {
          "type": "string"
        },
        "userName": {
          "type": "string"
        },
        "userFirstname": {
          "type": "string"
        }
      }
    },
    "v1GetUserInfoByUsernameResponse": {
      "type": "object",
      "properties": {
        "error": {
          "$ref": "#/definitions/v1Error"
        },
        "result": {
          "$ref": "#/definitions/v1GetUserInfoByUsernameResult"
        }
      }
    },
    "v1GetUserInfoByUsernameResult": {
      "type": "object",
      "properties": {
        "userId": {
          "type": "string"
        },
        "userName": {
          "type": "string"
        },
        "userFirstname": {
          "type": "string"
        }
      }
    }
  }
}